Moves task scheduler to package scheduler
