+++ benchflow-test-manager/application/src/main/java/cloud/benchflow/testmanager/resources/BenchFlowExperimentResource.java
@@ -11,7 +11,7 @@ import cloud.benchflow.testmanager.models.BenchFlowExperimentModel;
 import cloud.benchflow.testmanager.models.BenchFlowTestModel;
 import cloud.benchflow.testmanager.services.internal.dao.BenchFlowExperimentModelDAO;
 import cloud.benchflow.testmanager.services.internal.dao.BenchFlowTestModelDAO;
-import cloud.benchflow.testmanager.tasks.TestTaskScheduler;
+import cloud.benchflow.testmanager.scheduler.TestTaskScheduler;
 import io.swagger.annotations.Api;
 
 import javax.validation.Valid;
