+++ benchflow-test-manager/application/src/main/java/cloud/benchflow/testmanager/scheduler/TestTaskScheduler.java
@@ -1,4 +1,4 @@
-package cloud.benchflow.testmanager.tasks;
+package cloud.benchflow.testmanager.scheduler;
 
 import static cloud.benchflow.testmanager.models.BenchFlowTestModel.TestTerminatedState.GOAL_REACHED;
 
@@ -31,8 +31,7 @@ import org.slf4j.LoggerFactory;
  */
 public class TestTaskScheduler {
 
-  private static Logger logger =
-      LoggerFactory.getLogger(TestTaskScheduler.class.getSimpleName());
+  private static Logger logger = LoggerFactory.getLogger(TestTaskScheduler.class.getSimpleName());
 
   private ConcurrentMap<String, Future> testTasks = new ConcurrentHashMap<>();
