+++ benchflow-test-manager/application/src/main/java/cloud/benchflow/testmanager/tasks/running/DetermineExplorationStrategyTask.java
@@ -4,7 +4,7 @@ import cloud.benchflow.testmanager.BenchFlowTestManagerApplication;
 import cloud.benchflow.testmanager.exceptions.BenchFlowTestIDDoesNotExistException;
 import cloud.benchflow.testmanager.services.internal.dao.ExplorationModelDAO;
 import cloud.benchflow.testmanager.strategy.selection.ExperimentSelectionStrategy;
-import cloud.benchflow.testmanager.tasks.TestTaskScheduler;
+import cloud.benchflow.testmanager.scheduler.TestTaskScheduler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
